just deleted some files and cleaned up

+++ libft/src/ft_str/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: chrlomba <chrlomba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 14:33:09 by chrlomba          #+#    #+#             */
-/*   Updated: 2023/10/18 18:03:46 by chrlomba         ###   ########.fr       */
+/*   Updated: 2023/10/18 20:35:01 by chrlomba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int dstsize)
+size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
 	unsigned int	i;
 	unsigned int	j;
